--- src/pages/post/index.tsx
@@ -1,19 +1,0 @@
-import { CommentProps } from 'antd';
-import Post from '@components/Post';
-
-export interface PostProps {
-  editorName: string;
-  content: string;
-  tags?: string[];
-  comment?: CommentProps;
-}
-
-const PostPage: React.FC<PostProps> = () => {
-  return (
-    <>
-      <Post />
-    </>
-  );
-};
-
-export default PostPage;
